Improving UI and resource clean up

+++ app/src/main/java/com/mysamples/paginglistmovies/feature/model/TvShow.kt
@@ -5,7 +5,7 @@ data class TvShow(
     val originalName: String,
     val id: Int,
     val name: String,
-    val voteCount: Float,
+    val voteCount: Int,
     val voteAverage: Float,
     val imageUrl: String,
     val overview: String
